git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@433 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.Common/DataTransferObjects/AppForeclosureCaseSearchResultCollection.cs
@@ -7,6 +7,6 @@ namespace HPF.FutureState.Common.DataTransferObjects
 {
     public class AppForeclosureCaseSearchResultDTOCollection:BaseDTOCollection<AppForeclosureCaseSearchResultDTO>
     {
-        public int SearchResultCount { get; set; }
+        public double SearchResultCount { get; set; }
     }
 }
